Refactoring github driver (remove config provider)
